+++ packages/vda-did-resolver/src/index.ts
@@ -1,5 +1,7 @@
 import { getResolver } from './resolver'
+import { RPC_URLS } from './config'
 
 export {
     getResolver,
+    RPC_URLS
 }