+++ packages/vda-did-resolver/CHANGELOG.md
@@ -1,4 +1,5 @@
 
+- Support fetching default RPC_URLs
 - Add usage and configuration instructions to README.md
 
 2022-02-09 (v2.0.5)